refactor(state-migration): split in-process cluster fixture into separate files
