+++ Egil.Orleans.StateMigration/test/Egil.Orleans.StateMigration.Tests/OrleansInProcessClusterCollection.cs
@@ -1,0 +1,7 @@
+namespace Egil.Orleans.StateMigration.Tests;
+
+[CollectionDefinition(Name)]
+public sealed class OrleansInProcessClusterCollection : ICollectionFixture<OrleansInProcessClusterFixture>
+{
+    public const string Name = "OrleansInProcessCluster";
+}
