+++ Egil.Orleans.StateMigration/test/Egil.Orleans.StateMigration.Tests/StorageJsonOrleansInProcessTests.cs
@@ -1,7 +1,6 @@
 using System.Collections.Concurrent;
 using System.Text.Json;
 using Orleans.Storage;
-using Orleans.TestingHost;
 
 namespace Egil.Orleans.StateMigration.Tests;
 
@@ -73,64 +72,6 @@ public sealed class StorageJsonOrleansInProcessTests(OrleansInProcessClusterFixt
 
     private static string CreateToken(string prefix)
         => $"{prefix}-{Guid.NewGuid():N}";
-}
-
-[CollectionDefinition(Name)]
-public sealed class OrleansInProcessClusterCollection : ICollectionFixture<OrleansInProcessClusterFixture>
-{
-    public const string Name = "OrleansInProcessCluster";
-}
-
-public sealed class OrleansInProcessClusterFixture : IAsyncLifetime
-{
-    public InProcessTestCluster Cluster { get; private set; } = null!;
-
-    public SystemTextJsonGrainStorageSerializer DefaultSerializer { get; } = CreateDefaultSerializer();
-    public SystemTextJsonGrainStorageSerializer CustomSerializer { get; } = CreateCustomSerializer();
-
-    public async ValueTask DisposeAsync()
-    {
-        if (Cluster is not null)
-        {
-            await Cluster.DisposeAsync();
-        }
-    }
-
-    public async ValueTask InitializeAsync()
-    {
-        var builder = new InProcessTestClusterBuilder();
-        builder.ConfigureSilo((_, siloBuilder) =>
-        {
-            siloBuilder.AddMemoryGrainStorage("storage-default", optionsBuilder =>
-            {
-                optionsBuilder.Configure(options => options.GrainStorageSerializer = DefaultSerializer);
-            });
-
-            siloBuilder.AddMemoryGrainStorage("storage-custom", optionsBuilder =>
-            {
-                optionsBuilder.Configure(options => options.GrainStorageSerializer = CustomSerializer);
-            });
-        });
-
-        InProcessTestCluster cluster = builder.Build();
-        await cluster.DeployAsync();
-        Cluster = cluster;
-    }
-
-    private static SystemTextJsonGrainStorageSerializer CreateDefaultSerializer()
-    {
-        JsonSerializerOptions readAndWriteOptions = new JsonSerializerOptions().AddStateMigrationSupport();
-        JsonSerializerOptions flattenedWriteOptions = new JsonSerializerOptions()
-            .AddStateMigrationSupport(StoragePayloadLayout.Flattened);
-        return new SystemTextJsonGrainStorageSerializer(readAndWriteOptions, flattenedWriteOptions);
-    }
-
-    private static SystemTextJsonGrainStorageSerializer CreateCustomSerializer()
-    {
-        JsonSerializerOptions customOptions = new JsonSerializerOptions()
-            .AddStateMigrationSupport("_type", "_value");
-        return new SystemTextJsonGrainStorageSerializer(customOptions, flattenedWriteOptions: null);
-    }
 }
 
 internal interface IDefaultStorageStateGrain : IGrainWithStringKey
